Fixed after-load test setup inappropriate for other tests causing false negative.
